see full decription
The form is at the top of the site. Divided into two columns. From the top of the form, a drop-down list of topics is displayed in full width.
Left column: Surname, First name, Patronymic name, Email. mail, phone
Right column: Subject, Text of the appeal
At the bottom of the form is a checkbox, a button to send and a request history button.

Description of checks:
1. Surname, First name, Patronymic name, Email address. mail:
 HTML: Using the required attribute.
 CSS: Styling - width 100%, dark background, light text, dark border.
2. Telephone:
 HTML: Using the pattern="^\+380\d{9}$" attribute.
 CSS: Styling similar to other fields, can be empty.
3. Topic, Address:
 HTML: Using the required attribute.
 CSS: Styling similar to other fields, fixed size.
4. Topic:
 HTML: Using the required attribute.
 CSS: Styling - width 100%, dark background, light text, dark border.
5. Checkbox "Consent to the processing of personal data":
 HTML: Using the required attribute.
 CSS: Stylization similar to other fields, indicating consent to personal data processing.
6. "Send" button:
 HTML: Using the <button> tag.
 CSS: Styling - wide, dark green color, white text, underlined cursor.
Indication of invalid and valid data (at the HTML level):
 Fields with erroneous data are highlighted with a red frame, or an error message is displayed next to the field.
 Empty fields in which data must be entered are orange.
 Fields with valid data are highlighted with a green frame.

Currently, the functionality of the history button and checkbox is not implemented, it will be added in future works.