new validation
All validation is moved to js.
Only numbers can be entered in the number field,
and the autocomplete works without any problems.
The functionality of local history storage has been added.
The last three entries are saved.
The last record is erased when a new one is added.
Duplicate check available.