changelog
The history of requests is stored in an external xml file. Despite the fact that validation takes place on the client side (for speed, convenience, visualization, etc.), part of the validation is also implemented in the server part.
The server part is written in php.

A simplified description of the data saving process:
The JavaScript code in script.js initiates a POST request to the save_history.php file when the user submits the form.
Data from the form is serialized into JSON and sent in the request body.
A PHP script (save_history.php) on the server receives the JSON data using file_get_contents('php://input').
The PHP script processes the data, checks for duplicates, adds a new entry to the history.xml file, and saves it.
The PHP script sends a JSON response back to the client indicating whether the operation was successful or an error occurred.

Validation and error handling in the server

The script starts by reading the existing history data from the history.xml file using simplexml_load_file($historyFilePath).

Processing a GET request to retrieve history:
If the server receives a GET request, it checks whether the XML file is missing ($historyData === null).
If the XML file is missing, the script returns a 505 Internal Server Error.
If the XML file is present but empty ($historyData->count() === 0), the script returns a 204 No Content response.

Processing a POST request to save a new entry:
For a POST request, the script reads the input data from the request body using json_decode(file_get_contents('php://input'), true).
It checks if any of the required fields are empty, and if so, returns a 400 Bad Request response with an error message.
The script then checks for duplicate records by comparing the new data to the existing records.
If a duplicate is found, it returns an error(duplicate) response.
If there are already 10 entries, the oldest entry will be deleted to keep the history down to 10.
It adds a new record to the XML data, including the timestamp and any additional fields.
The updated XML data is stored in the history.xml file.

Error handling and success:
If any errors occur during the process (eg empty records, duplicate records), the script returns a JSON response with the error message and the corresponding HTTP response code.
If the operation is successful, the script returns a JSON response with a success message.

Also, added a session to the server part. It stores the time spent by the user to fill out the form and transfers its value to the xml after processing.
All data is stored in the xml file, but only the main ones are displayed when requested by the user. Retrieving data from XML is mainly handled by the jQuery library.
Additionally, the code has been reformatted and commented.
In case of a problem reading the API, a warning about this now appears under the time and date display on the form.